Refactor to go from nested schema to nested resource.
