+++ drowsy/resource.py
@@ -11,14 +11,10 @@
 from marshmallow.compat import with_metaclass
 from mqlalchemy import InvalidMQLException
 from sqlalchemy.exc import SQLAlchemyError
-import sys
 from drowsy import resource_class_registry
-from drowsy.fields import EmbeddableMixinABC, NestedRelated
+from drowsy.base import BaseResourceABC
 from drowsy.query_builder import QueryBuilder, SortInfo
-from drowsy.utils import get_error_message, get_field_by_dump_name
-from drowsy.exc import (
-    BadRequestError, UnprocessableEntityError, MethodNotAllowedError,
-    ResourceNotFoundError, MISSING_ERROR_MESSAGE)
+from drowsy.exc import BadRequestError
 
 
 class ResourceCollection(list):
@@ -46,163 +42,6 @@ class ResourceCollection(list):
 
         """
         return self.__len__()
-
-
-class ResourceABC(object):
-
-    """Abstract resource base class."""
-
-    def get(self, ident):
-        """Get an instance of this resource.
-
-        :param ident: Identifying info for the resource.
-        :return: The resource itself if found.
-        :raise ResourceNotFoundError: If no such resource exists.
-
-        """
-        raise NotImplementedError
-
-    def post(self, data):
-        """Create a resource with the supplied data.
-
-        :param data: Data used to create the resource.
-        :raise UnprocessableEntityError: If the supplied data cannot be
-            processed.
-        :return: The created resource.
-
-        """
-        raise NotImplementedError
-
-    def put(self, ident, data):
-        """Replace the identified resource with the supplied one.
-
-        :param ident: Identifying info for the resource.
-        :param data: Data used to replace the resource.
-        :raise ResourceNotFoundError: If no such resource exists.
-        :raise UnprocessableEntityError: If the supplied data cannot be
-            processed.
-        :return: The replaced resource.
-
-        """
-        raise NotImplementedError
-
-    def patch(self, ident, data):
-        """Update the identified resource with the supplied data.
-
-        :param ident: Identifying info for the resource.
-        :param data: Data used to update the resource.
-        :raise ResourceNotFoundError: If no such resource exists.
-        :raise UnprocessableEntityError: If the supplied data cannot be
-            processed.
-        :return: The updated resource.
-
-        """
-        raise NotImplementedError
-
-    def delete(self, ident):
-        """Delete the identified resource.
-
-        :param ident: Identifying info for the resource.
-        :raise ResourceNotFoundError: If no such resource exists.
-        :return: ``None``
-
-        """
-        raise NotImplementedError
-
-    def get_collection(self):
-        """Get a collection of resources."""
-        raise NotImplementedError
-
-    def post_collection(self, data):
-        """Create multiple resources in the collection of resources.
-
-        :param data: Data used to create the collection of resources.
-        :raise UnprocessableEntityError: If the supplied data cannot be
-            processed.
-        :return: ``None``
-
-        """
-        raise NotImplementedError
-
-    def put_collection(self, data):
-        """Replace the entire collection of resources.
-
-        :param data: Data used to replace the collection of resources.
-        :raise UnprocessableEntityError: If the supplied data cannot be
-            processed.
-        :return: ``None``
-
-        """
-        raise NotImplementedError
-
-    def patch_collection(self, data):
-        """Update the collection of resources.
-
-        :param data: Data used to update the collection of resources.
-        :raise UnprocessableEntityError: If the supplied data cannot be
-            processed.
-        :return: ``None``
-
-        """
-        raise NotImplementedError
-
-    def delete_collection(self):
-        """Delete all members of the collection of resources."""
-        raise NotImplementedError
-
-
-class SchemaResourceABC(ResourceABC):
-
-    """Abstract schema based resource class."""
-
-    def _get_schema_kwargs(self, schema_cls):
-        """Get default kwargs for any new schema creation.
-
-        :param schema_cls: The class of the schema being created.
-
-        """
-        raise NotImplementedError
-
-    def make_schema(self, subfilters=None, fields=None, embeds=None,
-                    partial=False, instance=None, strict=True):
-        """Used to generate a schema for this request.
-
-        :param subfilters: A dict of MQLAlchemy filters, with each key
-            being the dot notation of the relationship they are to be
-            applied to.
-        :type subfilters: dict or None
-        :param fields: Names of fields to be included in the result.
-        :type fields: list or None
-        :param embeds: A list of child resources (and/or their fields)
-            to include in the response.
-        :type embeds: list or None
-        :param bool partial: Whether partial deserialization is allowed.
-        :param instance: Object to associate with the schema.
-        :param bool strict: If ``True``, will raise an exception when
-            bad parameters are passed. If ``False``, will quietly ignore
-            any bad input and treat it as if none was provided.
-        :raise BadRequestError: Invalid fields or embeds will result
-            in a raised exception if strict is ``True``.
-        :return: A schema with the supplied fields and embeds included.
-        :rtype: :class:`~drowsy.schema.SchemaResourceABC`
-
-        """
-        raise NotImplementedError
-
-
-class NestableResourceABC(ResourceABC):
-
-    """Abstract nestable resource class."""
-
-    def make_subresource(self, name):
-        """Given a subresource name, construct a subresource.
-
-        :param str name: Dumped name of field containing a subresource.
-        :raise ValueError: If the name given isn't a valid subresource.
-        :returns: A constructed :class:`~drowsy.resource.Resource`
-
-        """
-        raise NotImplementedError
 
 
 class ResourceOpts(object):
@@ -294,484 +133,10 @@ class ResourceMeta(type):
         resource_class_registry.register(name, cls)
 
 
-class BaseResourceABC(SchemaResourceABC, NestableResourceABC):
-
-    OPTIONS_CLASS = ResourceOpts
-    default_error_messages = {
-        "validation_failure": "Unable to process entity.",
-        "invalid_embed": "Invalid embed supplied: %(embed)s",
-        "invalid_embeds": "Invalid embed supplied: %(embeds)s",
-        "invalid_field": "Invalid field supplied: %(field)s",
-        "invalid_fields": "Invalid fields supplied: %(fields)s",
-        "invalid_filters": "Invalid filters supplied.",
-        "commit_failure": "Unable to save the provided data.",
-        "invalid_collection_input": "The provided input must be a list.",
-        "resource_not_found": ("No resource matching the provided "
-                               "identity could be found."),
-        "invalid_sorts_type": "The sorts provided must be a list.",
-        "invalid_sort_type": "The sort provided is invalid.",
-        "invalid_sort_field": ("The sort provided for field %(field)s "
-                               "is invalid."),
-        "invalid_limit_value": ("The limit provided (%(limit)s) is not a "
-                                "non negative integer."),
-        "limit_too_high": ("The limit provided (%(limit)d) is greater than "
-                           "the max page size allowed (%(max_page_size)d)."),
-        "invalid_offset_value": ("The offset provided (%(offset)s) is not a "
-                                 "non negative integer."),
-        "method_not_allowed": ("The method (%(method)s) used to make this "
-                               "request is not allowed for this resource."),
-        "invalid_subresource_options": ("Limit and offset for this "
-                                        "subresource are not supported: "
-                                        "%(subresource_key)s"),
-        "invalid_subresource_filters": ("The supplied filters for this "
-                                        "subresource are invalid: "
-                                        "%(subresource_key)s"),
-        "invalid_subresource": "%(subresource_key)s is not a subresource.",
-        "invalid_subresource_limit": ("The limit (%(supplied_limit)s) given "
-                                      "for %(subresource_key)s subresource is "
-                                      "too high. The max limit is "
-                                      "%(max_limit)s."),
-        "invalid_subresource_sorts": ("The subresource %(subresource_key)s "
-                                      "can not have sorts applied without "
-                                      "a limit or offset being supplied.")
-    }
-
-    class Meta(object):
-        """Options object for a Resource.
-
-        Example usage: ::
-
-            class Meta:
-                schema_cls = MyModelSchema
-                error_messages = {
-                    "validation_failure": "Fix your data."
-                }
-
-        Available options:
-
-        - ``schema_cls``: The model schema this resource is built
-          around.
-        - ``error_messages``: A dict mapping an error message type to a
-          string or callable.
-
-        """
-
-    def __init__(self, context=None, page_max_size=None,
-                 error_messages=None, *args, **kwargs):
-        """Creates a new instance of the model.
-
-        :param context: Context used to alter the schema used for this
-            resource. For example, may contain the current
-            authorization status of the current request. If errors
-            should be translated, context should include a ``"gettext"``
-            key referencing a callable that takes in a string and any
-            keyword args.
-        :type context: dict, callable, or None
-        :param page_max_size: Used to determine the maximum number of
-            results to return by :meth:`get_collection`. Defaults to
-            the ``page_max_size`` from the resource's ``opts`` if
-            ``None`` is passed in. To explicitly allow no limit,
-            pass in ``0``. If given a ``callable``, it should accept
-            the resource itself as its first and only argument.
-        :type page_max_size: int, callable, or None
-        :param error_messages: May optionally be provided to override
-            the default error messages for this resource.
-        :type error_messages: dict or None
-
-        """
-        self._page_max_size = page_max_size
-        if self._page_max_size is None and hasattr(self.opts, "page_max_size"):
-            self._page_max_size = self.opts.page_max_size
-        self._context = context
-        # Set up error messages
-        messages = {}
-        for cls in reversed(self.__class__.__mro__):
-            messages.update(getattr(cls, 'default_error_messages', {}))
-        if isinstance(self.opts.error_messages, dict):
-            messages.update(self.opts.error_messages)
-        messages.update(error_messages or {})
-        self.error_messages = messages
-
-    def make_schema(self, fields=None, subfilters=None, embeds=None,
-                    partial=False, instance=None, strict=True):
-        """Used to generate a schema for this request.
-
-        :param fields: Names of fields to be included in the result.
-        :type fields: list or None
-        :param subfilters: Filters to be applied to children of this
-            resource. Each key in the dictionary should be a dot
-            notation key corresponding to a subfilter.
-        :type subfilters: dict or None
-        :param embeds: A list of relationship and relationship field
-            names to be included in the result.
-        :type embeds: list or None
-        :param bool partial: Whether partial deserialization is allowed.
-        :param instance: Object instance to associate with the schema.
-        :param bool strict: If ``True``, will raise an exception when
-            bad parameters are passed. If ``False``, will quietly ignore
-            any bad input and treat it as if none was provided.
-        :raise BadRequestError: Invalid fields or embeds will result
-            in a raised exception if strict is ``True``.
-        :return: A schema with the supplied fields and embeds included.
-        :rtype: :class:`~drowsy.schema.ResourceSchema`
-
-        """
-        # parse embed information
-        # combine subfilters with embeds to take care of implied embeds
-        if isinstance(embeds, list):
-            embeds_subfilters = [key for key in embeds]
-            if isinstance(subfilters, dict):
-                for key in subfilters:
-                    for embed in embeds:
-                        if key in embed:
-                            break
-                    else:
-                        embeds_subfilters.append(key)
-        elif isinstance(subfilters, dict):
-            embeds_subfilters = [key for key in subfilters]
-        else:
-            embeds_subfilters = None
-        converted_embeds, embed_name_mapping, embed_fields = (
-            self._get_embed_info(embeds=embeds_subfilters, strict=strict))
-        # fields
-        converted_fields = []
-        if fields:
-            for field in fields:
-                converted_field = self.convert_key_name(field)
-                if converted_field is None:
-                    if strict:
-                        self.fail("invalid_field", field=field)
-                elif converted_field:
-                    converted_fields.append(converted_field)
-        if converted_fields:
-            for embed_field in embed_fields:
-                if embed_field not in converted_fields:
-                    converted_fields.append(embed_field)
-            kwargs = self._get_schema_kwargs(self.schema_cls)
-            kwargs.update(
-                only=tuple(converted_fields),
-                partial=partial,
-                instance=instance
-            )
-            schema = self.schema_cls(**kwargs)
-        else:
-            kwargs = self._get_schema_kwargs(self.schema_cls)
-            kwargs.update(
-                partial=partial,
-                instance=instance
-            )
-            schema = self.schema_cls(**kwargs)
-        # actually attempt to embed now
-        for converted_embed in converted_embeds:
-            try:
-                schema.embed([converted_embed])
-            except AttributeError:
-                if strict:
-                    self.fail("invalid_embed",
-                              embed=embed_name_mapping[converted_embed])
-        return schema
-
-    def make_subresource(self, name):
-        """Given a subresource name, construct a subresource.
-
-        :param str name: Dumped name of field containing a subresource.
-        :raise ValueError: If the name given isn't a valid subresource.
-        :returns: A constructed :class:`~drowsy.resource.Resource`
-
-        """
-        # NOTE: No risk of BadRequestError here due to no embeds or
-        # fields being passed to make_schema
-        field = get_field_by_dump_name(self.make_schema(), dump_name=name)
-        if isinstance(field, NestedRelated):
-            return field.resource_cls(
-                **self._get_resource_kwargs(field.resource_cls)
-            )
-        raise ValueError("The provided name is not a valid subresource.")
-
-    def fail(self, key, errors=None, exc=None, **kwargs):
-        """Raises an exception based on the ``key`` provided.
-
-        :param str key: Failure type, used to choose an error message.
-        :param errors: May be used by the raised exception.
-        :type errors: dict or None
-        :param exc: If another exception triggered this failure, it may
-            be provided for a more informative failure.
-        :type exc: :exc:`Exception` or None
-        :param kwargs: Any additional arguments that may be used for
-            generating an error message.
-        :raise UnprocessableEntityError: If ``key`` is
-            ``"validation_failure"``. Note that in this case, errors
-            should preferably be provided.
-        :raise BadRequestError: The default error type raised in all
-            other cases.
-
-        """
-        unproccessable_errors = {"validation_failure", "commit_failure",
-                                 "invalid_collection_input"}
-        if key in unproccessable_errors:
-            raise UnprocessableEntityError(
-                code=key,
-                message=self._get_error_message(key, **kwargs),
-                errors=errors or {},
-                **kwargs)
-        elif key == "resource_not_found":
-            raise ResourceNotFoundError(
-                code=key,
-                message=self._get_error_message(key, **kwargs),
-                **kwargs
-            )
-        elif key == "method_not_allowed":
-            raise MethodNotAllowedError(
-                code=key,
-                message=self._get_error_message(key, **kwargs),
-                **kwargs)
-        else:
-            raise BadRequestError(
-                code=key,
-                message=self._get_error_message(key, **kwargs),
-                **kwargs)
-
-    def _get_error_message(self, key, **kwargs):
-        """Get an error message based on a key name.
-
-        If the error message is a callable, kwargs are passed
-        to that callable.
-
-        If ``self.context`` has a ``"gettext" key set to a callable,
-        that callable will be passed the resulting string and any
-        key word args for the sake of translation.
-
-        :param str key: Key used to access the error messages dict.
-        :param dict kwargs: Any additional arguments that may be passed
-            to a callable error message, or used to translate and/or
-            format an error message string.
-        :raise AssertionError: If ``key`` does not exist in the
-            error messages dict.
-        :return: An error message as a string.
-        :rtype: str
-
-        """
-        try:
-            return get_error_message(
-                error_messages=self.error_messages,
-                key=key,
-                gettext=self.context.get("gettext", None),
-                **kwargs)
-        except KeyError:
-            class_name = self.__class__.__name__
-            msg = MISSING_ERROR_MESSAGE.format(class_name=class_name, key=key)
-            raise AssertionError(msg)
-
-    @property
-    def schema_cls(self):
-        """Get the schema class associated with this resource.
-
-        :return: The class of the schema associated with this resource.
-
-        """
-        return self.opts.schema_cls
-
-    def whitelist(self, key):
-        """Determine whether a field is valid to be queried.
-
-        Uses the load_only property for the resource's schema fields
-        to determine whether the field should be queryable. Also handles
-        nested queries with the same logic.
-
-        :param str key: Dot notation field name. For example, if trying
-            to query an album, this may look something like
-            ``"tracks.playlists.track_id"``.
-
-        """
-        split_keys = key.split(".")
-        schema = self.schema_cls(**self._get_schema_kwargs(self.schema_cls))
-        while split_keys:
-            key = split_keys.pop(0)
-            if key in schema.fields:
-                field = schema.fields[key]
-                if field.load_only:
-                    return False
-                elif not split_keys:
-                    return True
-                elif isinstance(field, EmbeddableMixinABC):
-                    schema.embed([key])
-                if isinstance(field, NestedRelated):
-                    try:
-                        # try to use a subresource for whitelisting
-                        subresource = self.make_subresource(
-                            field.dump_to or field.name)
-                        if isinstance(subresource, BaseResourceABC):
-                            return subresource.whitelist(
-                                ".".join(split_keys))
-                        raise ValueError
-                    except (ValueError, TypeError):
-                        # if unable to do so, continue on iteratively
-                        # using this resource's whitelist rules
-                        schema = field.schema
-            else:
-                return False
-        return True
-
-    def convert_key_name(self, key):
-        """Given a dumped key name, convert to the name of the field.
-
-        :param str key: Name of the field as it was serialized, using
-            dot notation for nested fields.
-        :return: The key converted from it's dump form to its
-            internally used form.
-
-        """
-        schema = self.schema_cls(**self._get_schema_kwargs(self.schema_cls))
-        split_keys = key.split(".")
-        result_keys = []
-        while split_keys:
-            key = split_keys.pop(0)
-            field = get_field_by_dump_name(schema, key)
-            if field is not None:
-                result_keys.append(field.name)
-                if isinstance(field, EmbeddableMixinABC):
-                    schema.embed([key])
-                if isinstance(field, NestedRelated):
-                    try:
-                        # try to use a subresource for converting
-                        subresource = self.make_subresource(field.dump_to)
-                        if isinstance(subresource, BaseResourceABC):
-                            result_keys += subresource.convert_key_name(
-                                ".".join(split_keys)
-                            ).split(".")
-                            return ".".join(result_keys)
-                        raise ValueError
-                    except ValueError:
-                        # if unable to do so, continue on iteratively
-                        # using this resource's convert rules
-                        schema = field.schema
-                elif not split_keys:
-                    return ".".join(result_keys)
-                else:
-                    return None
-            else:
-                # Invalid key name, no matching field found.
-                return None
-        return ".".join(result_keys)
-
-    @property
-    def page_max_size(self):
-        """Get the max number of resources to return.
-
-        :return: The maximum number of resources to be
-            included in a result.
-        :rtype: int or None
-
-        """
-        if callable(self._page_max_size):
-            if sys.version_info.major == 2:  # pragma: no cover
-                func = self._page_max_size.__func__
-            else:
-                func = self._page_max_size
-            return func(self)
-        elif self._page_max_size is not None:
-            if self._page_max_size == 0:
-                return None
-            else:
-                return self._page_max_size
-
-    @property
-    def context(self):
-        """Return the schema context for this resource.
-
-        :return: A dictionary containing any context info used for this
-            resource.
-        :rtype: dict
-
-        """
-        if callable(self._context):
-            return self._context()
-        else:
-            if self._context is None:
-                self._context = {}
-            return self._context
-
-    @context.setter
-    def context(self, val):
-        """Set context to the provided value.
-
-        :param val: Used to set the current context value.
-        :type val: dict, callable, or None
-
-        """
-        self._context = val
-
-    def _get_schema_kwargs(self, schema_cls):
-        """Get default kwargs for any new schema creation.
-
-        :param schema_cls: The class of the schema being created.
-        :return: A dictionary of keyword arguments to be used when
-            creating new schema instances.
-        :rtype: dict
-
-        """
-        return {
-            "context": self.context
-        }
-
-    def _get_resource_kwargs(self, resource_cls):
-        """Get default kwargs for any new sub resource creation.
-
-        :param resource_cls: The class of the resource being created.
-        :return: A dictionary of keyword arguments to be used when
-            creating child resources.
-        :rtype: dict
-
-        """
-        return {
-            "context": self.context
-        }
-
-    def _get_embed_info(self, embeds=None, strict=True):
-        """Helper function that handles the supplied embeds.
-
-        :param embeds: A list of relationship and relationship field
-            names to be included in the result.
-        :type embeds: list or None
-        :param bool strict: If ``True``, will raise an exception when
-            bad parameters are passed. If ``False``, will quietly ignore
-            any bad input and treat it as if none was provided.
-        :raise BadRequestError: If ``strict`` is ``True`` and ``embeds``
-            are not valid.
-        :return: A list of converted embed field names, a dict mapping
-            their original name to their converted name, and a list of
-            the top level embed fields to be included.
-        :rtype: tuple
-
-        """
-        # embed converting
-        # name mapping used purely for error purposes
-        # key is converted name, value is orig attr name
-        embed_name_mapping = {}
-        converted_embeds = []
-        embed_fields = set()
-        if embeds is None:
-            embeds = []
-        for embed in embeds:
-            converted_embed = self.convert_key_name(embed)
-            embed_name_mapping[converted_embed] = embed
-            if converted_embed is None:
-                if strict:
-                    self.fail("invalid_embed", embed=embed)
-            elif converted_embed:
-                # used so if a fields param is provided, embeds are
-                # still included.
-                # e.g. albums?fields=album_id,tracks.track_id
-                #             &embeds=tracks.title
-                # tracks.title will get added to fields to include.
-                embed_fields.add(converted_embed.split(".")[0])
-            converted_embeds.append(converted_embed)
-        return converted_embeds, embed_name_mapping, embed_fields
-
-
 class BaseModelResource(BaseResourceABC):
 
     """Model API Resources should inherit from this object."""
+    OPTIONS_CLASS = ResourceOpts
 
     def __init__(self, session, context=None, page_max_size=None,
                  error_messages=None):
